Search: Parse lat,lng string into float64 coordinates #4930 #4812

+++ internal/meta/json_exiftool.go
@@ -120,7 +120,7 @@ func (data *Data) Exiftool(jsonData []byte, originalName string) (err error) {
 
 				if f := jsonValue.Float(); f != 0 {
 					fieldValue.SetFloat(f)
-				} else if f = txt.Float(jsonValue.String()); f != 0 {
+				} else if f = txt.Float64(jsonValue.String()); f != 0 {
 					fieldValue.SetFloat(f)
 				}
 			case uint, uint64:
